+++ app/dashboard/months/components/MonthDrawerContent.tsx
